refactor(engine): overhaul python module structure and imports
Restructures the engine's Python modules to function as a proper package. This is achieved by moving the `config` module inside the `StreamStorm` directory and converting all absolute imports to relative ones. A `sys.path` modification in the root `__init__.py` ensures robust import resolution.

This change also includes several stability improvements:
- Increases Playwright timeouts to 45 seconds to prevent failures on slow connections.
- Adds a generic exception handler to the start endpoint to ensure the application state is properly reset on unexpected errors.
- Moves and refines the `is_instance_alive` logic for more accurate browser state detection.
