+++ .gitignore
@@ -4,3 +4,4 @@ output
 .env
 .env.production
 *.log
+diff.txt
