+++ Engine/StreamStorm/api/lib/StormStopper.py
@@ -4,6 +4,7 @@ from logging import getLogger, Logger
 
 from ...core.StreamStorm import StreamStorm
 
+
 logger: Logger = getLogger(f"fastapi.{__name__}")
 
 async def background_storm_stopper() -> NoReturn:
