+++ Engine/StreamStorm/config/config.py
@@ -1,0 +1,6 @@
+
+CONFIG: dict = {
+    "ENV": "development", # Valid values: ["development", "production", "test"]
+}
+
+__all__: list[str] = ["CONFIG"]
