--- Engine/config/config.py
@@ -1,6 +1,0 @@
-
-CONFIG: dict = {
-    "ENV": "production", # Valid values: ["development", "production", "test"]
-}
-
-__all__: list[str] = ["CONFIG"]
