+++ Engine/main.py
@@ -16,7 +16,7 @@ from webview import create_window, start
 
 from StreamStorm import app
 from StreamStorm.utils.CustomLogger import CustomLogger
-from config.config import CONFIG 
+from StreamStorm.config.config import CONFIG 
 
 CustomLogger().setup_streamstorm_logging()
 
@@ -48,6 +48,8 @@ def main() -> None:
 
         ui_url: str = "https://streamstorm-ui.web.app/" if CONFIG["ENV"] == "production" else "http://localhost:5173"
         # ui_url = "https://streamstorm-ui.web.app/" 
+        
+        logger.debug(f"UI URL: {ui_url}")
 
         create_window(
             title="StreamStorm",
